# backend/go/services/tenant-manager/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.APIKeyResponse:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      key_prefix:
        type: string
      last_used_at:
        type: string
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
      raw_key:
        description: Only returned on create
        type: string
      status:
        type: string
      tenant_id:
        type: string
    type: object
  handler.CreateAPIKeyRequest:
    properties:
      description:
        type: string
      expiry_days:
        type: integer
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
    type: object
...
  handler.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: Tenant and API key management service
  title: Tenant Manager API
  version: "1.0"
paths:
...
  /api/v1/tenants/{id}/api-keys:
    get:
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: Filter only active keys
        in: query
        name: active_only
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.APIKeyResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List API keys
      tags:
      - APIKeys
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: API key payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.APIKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create API key
      tags:
      - APIKeys
  /api/v1/tenants/{id}/api-keys/{keyId}:
    delete:
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: API key ID
        in: path
        name: keyId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete API key
      tags:
      - APIKeys
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer access token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"