# backend/go/services/tenant-manager/internal/adapter/http/handler/swagger_models.go
package handler

// ErrorResponse documents the error payload returned by the API.
type ErrorResponse struct {
	Code    string      `json:"code"`
	Message string      `json:"message"`
	Details interface{} `json:"details,omitempty"`
}